create dedicated chat component
